backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
langchain
openai
pymongo
motor
python-dotenv
firebase-admin
python-dateutil
pytest
pytest-asyncio
google-generativeai
python-jose[cryptography]
passlib[bcrypt]
python-multipart
